Optimize Marilyn Monroe

+++ catscripts/Entertainment/Actors_and_actresses/Marilyn_Monroe/MarilynMonroe.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 
-export KEYWORDS_ACTRESS_MARILYNMONROE="Marilyn(| )Monroe"
+KEYWORDS_ACTRESS_MARILYNMONROE="Marilyn(| )Monroe"
 
-egrep -i "$KEYWORDS_ACTRESS_MARILYNMONROE" newpages.txt >> MarilynMonroe.txt
-
-MARILYNMONROE=`stat --print=%s MarilynMonroe.txt`
-
-if [ $MARILYNMONROE -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="MarilynMonroe.txt"
-  export CATNAME="Marilyn Monroe"
-  $CATEGORIZE
-fi
 
-rm MarilynMonroe.txt
+  MARILYNMONROE=`egrep -i "$KEYWORDS_ACTRESS_MARILYNMONROE" newpages.txt`
+
+  if [ $MARILYNMONROE -ne 0 ];
+  then
+    egrep -i "$KEYWORDS_ACTRESS_MARILYNMONROE" newpages.txt >> MarilynMonroe.txt
+    export CATFILE="MarilynMonroe.txt"
+    export CATNAME="Marilyn Monroe"
+    $CATEGORIZE
+    rm MarilynMonroe.txt
+    unset MARILYNMONROE
+  fi
+fi
